# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "TW Company",
      "description": "TW Company список проектов и задач",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "10.100.100.62:3001",
    "basePath": "/",
    "tags": [
      {
        "name": "Progect",
        "description": "Работа с проектами"
      }
    ],
    "paths": {
        "/api/getAllProgect/": {
          "get": {
            "tags": ["Progect"],
            "summary": "Получение списка всех проктов",
            "responses": {
              "200": {
                "description": "OK",
                "schema": {
                  "$ref": "#/definitions/Users"
                }
              }
            }
          }
        },
        "/api/getProgectById/:id": {
          "get": {
            "tags": ["Progect"],
            "parameters": [
              {
                "name": "id",
                "in": "params",
                "description": "User that we want to create",
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            ],
            "summary": "Получение прокта по ID",
            "responses": {
              "200": {
                "description": "OK",
                "schema": {
                  "$ref": "#/definitions/Users"
                }
              }
            }
          }
        },
        "/addUser": {
          "post": {
            "tags": ["Progect"],
            "description": "Create new user in system",
            "parameters": [
              {
                "name": "user",
                "in": "body",
                "description": "User that we want to create",
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            ],
            "produces": ["application/json"],
            "responses": {
              "200": {
                "description": "New user is created",
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        }
    },
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"]
}